[Samsung] Deal with core-libart.jar's field variants
It's known that there are two more field in java.lang.DexCache class.
But some devices have only one more, or even don't have such fields.
So this quite dirty workaround was introduced.

The known fields are
  * String[] literals
  * Object z_padding

Workaround : Search for the fields in /system/framework/core-libart.jar when
             Android Runtime initialises then set up a flag whether to adjust
             offsets or not

I couldn't think of a better logic. It looks ugly but hopefully, seems to work fine.